+++ src/app/pages/home/home.component.ts
@@ -1,6 +1,6 @@
 import {Component, effect, ElementRef, inject, OnInit, signal, ViewChild} from '@angular/core';
 import {ModeService} from "../../services/mode.service";
-import {ButtonComponent} from '../../components/button/button.component';
+import {ButtonComponent} from '../../layout/button/button.component';
 import {TitleComponent} from '../../components/title/title.component';
 import {ActivityService} from '../../services/activity.service';
 import genUniqueId from '../../utils/genUniqueId';
